Fix indexExists() and clarify its usage - #42
Was always returning true because the groovy truthiness of a non-null
object is true.

+++ index/src/integrationTest/groovy/ncei/catalog/controller/IndexRabbitApiSpec.groovy
@@ -24,7 +24,8 @@ class IndexRabbitApiSpec extends Specification {
 
   def setup() {
     service.INDEX = 'test_index'
-    service.deleteIndex()
+    if (service.indexExists()) { service.deleteIndex() }
+    service.createIndex()
   }
 
   def 'rabbit save to elastic search works'() {
